Extract cmd and hex helpers to files
